long totalElementos

+++ backend/src/main/java/br/com/lacasa/sistemavendas/dto/PaginaResponseDTO.java
@@ -2,11 +2,14 @@ package br.com.lacasa.sistemavendas.dto;
 
 import java.util.List;
 
+
+
+
 public record PaginaResponseDTO<T>(
 		List<T> conteudo,//pode ser uma lista de clientes, produtos ou pedidos
 		int paginaAtual,
 		int tamanhoPagina,
-		int totalElementos,
+		long totalElementos,
 		int totalPaginas,
 		boolean primeira,
 		boolean ultima) {
